% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control2pan.R
\name{control2pan}
\alias{control2pan}
\title{La fonction "controlpan"}
\usage{
control2pan(pan1 = ech, pan2 = smple)
}
\arguments{
\item{pan1}{=ech, objepan 1}

\item{pan2}{=smple, objepan 2}

\item{var.control}{="Departement", variable de controle}

\item{vars.test}{=c("Anciennete", "rur"), variable de test}
}
\value{
listidfi, un df avec les prop
}
\description{
"control2pan" permet de comparer deux objets objepan
}
\references{
Voir basiques::objepan() et les packages "sampling" et "survey"; A creuser!
}
\author{
Elie
}
