% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlpan.R
\name{controlpan}
\alias{controlpan}
\title{La fonction "controlpan"}
\usage{
controlpan(data = datas, varpan = "Est.dans.l.echantillon.regional",
  var.control = "Departement", vars.test = c("Anciennete", "rur"),
  mod.keep.sample = c("REG", "Ensemble"))
}
\arguments{
\item{data}{=datas, data.frame de base}

\item{varpan}{="Est.dans.l.echantillon.regional", variable de panel}

\item{var.control}{="Departement", variable de controle}

\item{vars.test}{=c("Anciennete", "rur"), variable de test}
}
\value{
listidfi, un df avec les prop
}
\description{
"controlpan" permet de controler la repartition des modalites des variables vars.test, dans la sous-population precisee par var.control, selon les modalites de la variable de panel varpan
}
\references{
Voir les packages "sampling" et "survey"; A creuser!
}
\author{
Elie
}
