% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_to_R.R
\name{excel_to_R}
\alias{excel_to_R}
\title{La fonction "excel_to_R"}
\usage{
excel_to_R(autre.separate = FALSE, ...)
}
\arguments{
\item{autre.separate}{si FALSE, cela veut dire que dans le docx les "Autres précisez: ..." sont présentés comme une modalité de réponse parmi les autres}

\item{...}{le chemin du fichier, et autres arguments de la fonction officer::read_docx(...)}
}
\description{
excel_to_R utilise le package officer pour créer un df un peu lisible à partir d'une docx d'enquête
}
\author{
Elie
}
