% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objepan.R
\name{objepan}
\alias{objepan}
\title{La fonction "objepan"}
\usage{
objepan(Objectif.total = 400, data.frame.sample = datas.pannb,
  nb.criteres = 2, datas.crit = list(popudep.AURA, sample.type),
  type.crit = c("eff", "prop"), sample.index = c("Departement",
  "Instit.Assoc"), crit.index = c("Nom.du.departement", "type"),
  exclude.sample = c("0", "0"), controls = TRUE,
  controls.vars = c("Milieu"), poid.supp = FALSE,
  var.poid.supp = "var.poid.supp", hierarch.crit = TRUE,
  idvar.sample = NULL)
}
\arguments{
\item{Objectif.total}{400 # taille du sample}

\item{data.frame.sample}{datas.pannb #data.frame avec les donnees non samplees}

\item{nb.criteres}{2 nb de criteres externes de sample (sur la population de reference)}

\item{datas.crit}{list(popudep.AURA, sample.type) liste des data.frame avec les donnees externes, dans l'ordre}

\item{type.crit}{c("eff", "prop") pour chaque df externe, dans l'ordre: s'agit-il des effectifs ou des proportions?}

\item{sample.index}{c("Departement", "Instit.Assoc") pour chaque df externe, dans l'ordre: la variable correpsondante dans le df non sample}

\item{crit.index}{c("Nom.du.departement", "type") pour chaque df externe, dans l'ordre: la variable correspondante dans chaque df externe}

\item{exclude.sample}{c("0", "0") pour chaque df externe, dans l'ordre: levels non pertinents qui trainent}

\item{controls}{FALSE logical: des variables dans le df non sample serviront-elles de controle au moment du sample?}

\item{controls.vars}{c("Statut", "Milieu") nom (characters) des variables de controle}

\item{var.poid.supp}{"var.poid.supp"}

\item{hierarch.crit}{TRUE}

\item{idvar.sample}{NULL le cas echeant, nom (character) de la vairable d'id dans le df non sample}
}
\value{
res<-list("final.obj"=final.obj, "crit"=lista, "sample.size"=sample.size, "sample.id"=sample.id, "newdata"=listou.df, "newdata.list"=listou)
}
\description{
"objepan" permet de sampler un data.frame en recourant a des probabilites (statistiques) externes.
}
\note{
Typiquement: la repartition effective de la population par departements dans l'ensemble de la region AURA hors enquete. objepan permet aussi de sampler, a partir de ces prob externes, en prenant en compte la representativite interne de chaque sous-echantillon: il me fait 20 personnes de la Loire, mais dans mon data.frame il y a 50% d'hommes dans la loire, je veux donc selectionner dans  ces 20 personnes environ 10 femmes et 10 hommes.
}
\examples{
objepan<-function(Objectif.total = 400, data.frame.sample=datas.pannb, nb.criteres=2,
datas.crit=list(popudep.AURA, sample.type), type.crit=c("eff", "prop"),
sample.index=c("Departement", "Instit.Assoc"),
crit.index=c("Nom.du.departement", "type"),
exclude.sample=c("0", "0"),
controls=FALSE,
controls.vars=c("Statut", "Milieu"),
var.poid.supp="var.poid.supp",
hierarch.crit=TRUE,
idvar.sample=NULL)->test
}
\references{
Voir les packages "sampling" et "survey"; A creuser!
}
\author{
Elie
}
