% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objepan3.R
\name{objepan3}
\alias{objepan3}
\title{La fonction "objepan3"}
\usage{
objepan3(Objectif.total = 400, data.frame.sample = datas.pannb,
  nb.criteres = 2, datas.crit = list(popudep.AURA, sample.type),
  type.crit = c("eff", "prop"), sample.index = c("Departement",
  "Instit.Assoc"), crit.index = c("Nom.du.département", "type"),
  exclude.sample = c("0", "0"), varcontrol, nb.criteres2 = 2,
  sample.index2 = c("rur", "anclog"), exclude.sample2 = c("0", "0"),
  idvar.sample = NULL)
}
\arguments{
\item{Objectif.total}{k]] # taille du sample}

\item{data.frame.sample}{datas.pannb #data.frame avec les donnees non samplees}

\item{nb.criteres}{2 nb de criteres externes de sample (sur la population de reference)}

\item{datas.crit}{list(popudep.AURA, sample.type) liste des data.frame avec les donnees externes, dans l'ordre}

\item{type.crit}{c("eff", "prop") pour chaque df externe, dans l'ordre: s'agit-il des effectifs ou des proportions?}

\item{sample.index}{c("Departement", "Instit.Assoc") pour chaque df externe, dans l'ordre: la variable correpsondante dans le df non sample}

\item{crit.index}{c("Nom.du.departement", "type") pour chaque df externe, dans l'ordre: la variable correspondante dans chaque df externe}

\item{exclude.sample}{c("0", "0") pour chaque df externe, dans l'ordre: levels non pertinents qui trainent}

\item{varcontrol}{= c("rur", "anclog") Vecteur des noms des variables (factor) servant de contrôle
@param nb.criteres2 plus utilisé = 0, #2
@param sample.index2 plus utilisé= c("rur", "anclog"),
@param exclude.sample2 plus utilisé= c("0", "0"),
@param idvar.sample  NULL le cas echeant, nom (character) de la vairable d'id dans le df non sample}
}
\value{
res<-list("final.obj"=final.obj, "ensemble"=Ensemble, "crit"=lista, "sample.size"=sample.size, "sample.id"=sample.id,
"newdata"=listou3.df, "newdata.list"=listou2, "listec"=listec, "listec.df"=listec3,
"weighted.quant"=weighted.quant, "weighted.tab"=weighted.tab)
}
\description{
"objepan3" permet de sampler un data.frame en recourant a des probabilites (statistiques) externes.
}
\note{
Typiquement: la repartition effective de la population par departements dans l'ensemble de la region AURA hors enquete. objepan permet aussi de sampler, a partir de ces prob externes, en prenant en compte la representativite interne de chaque sous-echantillon: il me fait 20 personnes de la Loire, mais dans mon data.frame il y a 50% d'hommes dans la loire, je veux donc selectionner dans  ces 20 personnes environ 10 femmes et 10 hommes.
}
\examples{
objepan3(Objectif.total = dim(datas_e3)[1], data.frame.sample = datas_e3, nb.criteres = 2, datas.crit = list(popudep.AURA, sample.type), type.crit = c("eff", "prop"), sample.index = c("Departement", "Instit.Assoc"), crit.index = c("Nom.du.département", "type"), exclude.sample = c("0", "0") ,
         nb.criteres2 = 0, #2
         #sample.index2 = c("rur", "anclog"),
         #exclude.sample2 = c("0", "0"),
         idvar.sample = "token",
         varcontrol = c("rur", "anclog"))->objpan_e3
}
\references{
Voir les packages "sampling" et "survey"; A creuser!
}
\author{
Elie
}
