% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/que_to_tabvide.R
\name{que.to.tabvide}
\alias{que.to.tabvide}
\title{La fonction "que.to.tabvide"}
\usage{
que.to.tabvide(SPHINX_ou_WORD = "WORD", data = df$df.QR, ...)
}
\arguments{
\item{SPHINX_ou_WORD}{"WORD"}

\item{data}{=df$df.QR}
}
\description{
que.to.tabvide utilise le résultat de excel_to_R() pour créer une table vide reprenant la structure de 'lenquête.
}
\author{
Elie
}
