% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpss.R
\name{rpss}
\alias{rpss}
\title{La fonction "rpss"}
\usage{
rpss(stratum, n, probum = NULL, REPLACE)
}
\arguments{
\item{stratum}{factor variable sur laquelle échantilloner}

\item{n}{taille de l'échantillon   d}
}
\value{
Résultat de la fonction survey::stratsample. Indice dans stratum
}
\description{
function to take a Random Proportional Stratified Sample (RPSS) of size n
}
\examples{
voir objepan(). rpss(stratum=interaction(df$var1, df$var2), n=Obj)->samplec
}
\references{
Voir les packages "sampling" et "survey"; A creuser!
}
\author{
Jean V. Adams https://stackoverflow.com/questions/17002101/random-sample-with-multiple-probabilities-in-r
}
