% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_by_tokens.R
\name{url.by.tokens}
\alias{url.by.tokens}
\title{La fonction "url.by.tokens"}
\usage{

  url.by.tokens(text.before.token = " https://enqconj-mrie.limequery.org/128784?token=",
  text.after.token = "&lang=fr", data = contacts, token.var = NULL,
  email.var = NULL)
}
\arguments{
\item{text.before.token}{première partie de l'url avant le token}

\item{text.after.token}{deuxième partie de l'url après le token}

\item{data}{=contacts, data.frame avec au moins les tokens et les emails}

\item{token.var}{=NULL character string ou indice numérique. Si NULL alors tokenvar<-"token"}

\item{email.var}{=NULL character string ou indice numérique. Si NULL alors emailvar<-"email"}
}
\value{
res est une liste avec le df contenant les emails et les urls, et un data.frame de control avec les emails, les url et les tokens.
}
\description{
"url.by.tokens" permet de créer rapidement les URL d'accès aux questionnaires LimeSurvey pour les contacts invités
}
\note{
Voir sur LimeSurvey pour les infos d'URL et les tokens.
}
\examples{
url.by.tokens(text.before.token=" https://enqconj-mrie.limequery.org/128784?token=",
text.after.token="&lang=fr",
data=contacts,
token.var=NULL,
email.var=NULL)
}
\author{
Elie
}
